Default fuel-code -> base-fuel-type mapping (BC FUEL_TYPE_CD factor levels)

Returns the mapping used by gitanyow-partial-harvest and other projects that use the BC FUEL_TYPE_CD factor encoding for fuel_types_rast. Levels correspond to: 1=B71_S-2, 2=C-2, 3=C-3, 4=C-4, 5=C-5, 6=C-6, 7=C-7, 8=D-1/2, 9=M-1/2, 10=N (non-fuel), 11=O-1a/b, 12=S-1, 13=S-3. Mapped to the five base types accepted by defaultFuelTypeTable() / calibration_par_names().

Usage

bc_fuel_code_to_base()

Value

Character vector of length 13, names "1".."13", values "Conifer" / "ConiferPlantation" / "Deciduous" / "Slash" / "Open" / NA_character_.

Details

Downstream projects with a different fuel-classification raster should pass their own mapping vector to save_observed_fire_targets() via the fuel_code_to_base argument.

See also