Default PnET-Succession DisturbanceReductions table

Returns a sensible default data.frame for the PnET-Succession DisturbanceReductions input file with columns for the four most common disturbance types (fire, wind, harvest, bda). Values are taken from the example in 10.1 of the user guide and should be tuned to the study system. The first column (Pool) names the dead-pool reduction, drawn from .pnetDisturbancePools.

Usage

defaultPnETDisturbanceReductions()

Value

data.frame

See also